backend/requirements.txt
requests==2.32.3
beautifulsoup4==4.13.4
openai==1.65.4
psycopg2-binary==2.9.10
python-dotenv==1.1.0